% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio.boxplot.R
\name{bio.boxplot}
\alias{bio.boxplot}
\title{Data Visualization}
\usage{
bio.boxplot(filename = "", start, end, ref.name = "")
}
\arguments{
\item{input}{excel file path}
}
\description{
Data Visualization
}
\examples{
x axis is categorical, y axis is numeric box plot


}
\keyword{box}
\keyword{plot}
