% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio.jitterplot.R
\name{bio.jitterplot}
\alias{bio.jitterplot}
\title{Data Visualization}
\usage{
bio.jitterplot(filename = "", start = 2, end = 3, ref.name = "")
}
\arguments{
\item{input}{excel file path}
}
\description{
Data Visualization
}
\examples{
x axis is categorical, y axis is numeric box plot

}
\keyword{jitter}
\keyword{plot}
