% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio.tost.R
\name{bio.tost}
\alias{bio.tost}
\title{TOST Function}
\usage{
bio.tost(
  filename,
  sheet.no = 1,
  raw.data,
  test.name,
  ref.name,
  method.name,
  ci = 0.9,
  paired = FALSE,
  sample.n.show = TRUE
)
}
\arguments{
\item{The}{equivalence test for biosimilar and reference product.}
}
\description{
This function allows Equivalence test.
}
\examples{
biotest.tost(test.name="Biosimilar", ref.name="Reference", method.name="WAX-HPLC", ci=0.95, paired = FALSE )
test.name Write a test sample name.
ref.name Write a reference sample name.
method.name Write a test method name for title
ci Confidence interval
paired default is FALSE
}
\keyword{TOST}
\keyword{biosimilar}
